workaround for menu corruption with Adreno driver
From vkCmdBindDescriptorSets page:

"A compatible descriptor set must be bound for all set
numbers that any shaders in a pipeline access, at the time
that a draw or dispatch command is recorded to execute
using that pipeline. However, if none of the shaders in a
pipeline statically use any bindings with a particular set
number, then no descriptor set need be bound for that set
number, even if the pipeline layout includes a non-trivial
descriptor set layout for that set number."

The gp_color graphics pipeline only uses descriptor sets
(0,1) however they may be undefined when switching from a
graphics pipeline which has more than two descriptor sets.

This workaround ensures that all descriptor sets are bound
even when some are not used.